better separation of concerns in preset
